Finished Annotated call
